decompose some code from RHM into new class to clear things a bit
