+++ src/main/scala/beam/agentsim/agents/ridehail/RideHailManager.scala
@@ -1,6 +1,5 @@
 package beam.agentsim.agents.ridehail
 
-import java.util
 import java.util.concurrent.TimeUnit
 
 import akka.actor.{ActorLogging, ActorRef, Props, Stash}
@@ -14,13 +13,9 @@ import beam.agentsim.agents.BeamAgent.Finish
 import beam.agentsim.agents.modalbehaviors.DrivesVehicle._
 import beam.agentsim.agents.ridehail.RideHailAgent._
 import beam.agentsim.agents.ridehail.RideHailManager._
+import beam.agentsim.agents.ridehail.RideHailVehicleManager.{Available, InService, RideHailAgentLocation}
 import beam.agentsim.agents.ridehail.allocation._
-import beam.agentsim.agents.vehicles.AccessErrorCodes.{
-  CouldNotFindRouteToCustomer,
-  DriverNotFoundError,
-  RideHailVehicleTakenError
-}
-import beam.agentsim.agents.vehicles.BeamVehicle.BeamVehicleState
+import beam.agentsim.agents.vehicles.AccessErrorCodes.{CouldNotFindRouteToCustomer, DriverNotFoundError, RideHailVehicleTakenError}
 import beam.agentsim.agents.vehicles.VehicleProtocol.StreetVehicle
 import beam.agentsim.agents.vehicles.{PassengerSchedule, _}
 import beam.agentsim.events.SpaceTime
@@ -39,12 +34,8 @@ import com.google.common.cache.{Cache, CacheBuilder}
 import com.vividsolutions.jts.geom.Envelope
 import org.matsim.api.core.v01.population.Person
 import org.matsim.api.core.v01.{Coord, Id}
-import org.matsim.core.utils.collections.QuadTree
-import org.matsim.core.utils.geometry.CoordUtils
 import org.matsim.vehicles.Vehicle
 
-import scala.collection.JavaConverters._
-import scala.collection.immutable.HashSet
 import scala.collection.mutable
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -91,7 +82,6 @@ object RideHailManager {
     )
   }
 
-  sealed trait RideHailServiceStatus
 
   case class NotifyIterationEnds()
 
@@ -142,26 +132,10 @@ object RideHailManager {
 
   case class RegisterRideUnavailable(ref: ActorRef, location: Coord)
 
-  case class RideHailAgentLocation(
-    rideHailAgent: ActorRef,
-    vehicleId: Id[Vehicle],
-    currentLocationUTM: SpaceTime
-  ) {
-
-    def toStreetVehicle: StreetVehicle = {
-      StreetVehicle(vehicleId, currentLocationUTM, CAR, asDriver = true)
-    }
-  }
-
-  case class RideHailAgentETA(
-    agentLocation: RideHailAgentLocation,
-    distance: Double,
-    timeToCustomer: Double
-  )
 
   case class RepositionResponse(
     rnd1: RideHailAgentLocation,
-    rnd2: RideHailManager.RideHailAgentLocation,
+    rnd2: RideHailAgentLocation,
     rnd1Response: RoutingResponse,
     rnd2Response: RoutingResponse
   )
@@ -177,13 +151,6 @@ object RideHailManager {
   case object DebugRideHailManagerDuringExecution
 
   case class ContinueBufferedRideHailRequests(tick: Int)
-
-  /* Available means vehicle can be assigned to a new customer */
-  case object Available extends RideHailServiceStatus
-
-  case object InService extends RideHailServiceStatus
-
-  case object OutOfService extends RideHailServiceStatus
 
 }
 
@@ -205,6 +172,8 @@ class RideHailManager(
   /**
     * Customer inquiries awaiting reservation confirmation.
     */
+  val vehicleManager: RideHailVehicleManager = new RideHailVehicleManager(this, boundingBox)
+
   lazy val travelProposalCache: Cache[String, TravelProposal] = {
     CacheBuilder
       .newBuilder()
@@ -242,8 +211,6 @@ class RideHailManager(
       self,
       this
     )
-  private val vehicleState: mutable.Map[Id[Vehicle], BeamVehicleState] =
-    mutable.Map[Id[Vehicle], BeamVehicleState]()
   private val DefaultCostPerMinute = beamServices.beamConfig.beam.agentsim.agents.rideHail.defaultCostPerMinute
   tncIterationStats.foreach(_.logMap())
   private val DefaultCostPerSecond = DefaultCostPerMinute / 60.0d
@@ -251,33 +218,6 @@ class RideHailManager(
   beamServices.beamRouter ! GetTravelTime
   beamServices.beamRouter ! GetMatSimNetwork
   //TODO improve search to take into account time when available
-  private val availableRideHailAgentSpatialIndex = {
-    new QuadTree[RideHailAgentLocation](
-      boundingBox.getMinX,
-      boundingBox.getMinY,
-      boundingBox.getMaxX,
-      boundingBox.getMaxY
-    )
-  }
-  private val inServiceRideHailAgentSpatialIndex = {
-    new QuadTree[RideHailAgentLocation](
-      boundingBox.getMinX,
-      boundingBox.getMinY,
-      boundingBox.getMaxX,
-      boundingBox.getMaxY
-    )
-  }
-  private val outOfServiceRideHailAgentSpatialIndex = {
-    new QuadTree[RideHailAgentLocation](
-      boundingBox.getMinX,
-      boundingBox.getMinY,
-      boundingBox.getMaxX,
-      boundingBox.getMaxY
-    )
-  }
-  private val availableRideHailVehicles = mutable.HashMap[Id[Vehicle], RideHailAgentLocation]()
-  private val outOfServiceRideHailVehicles = mutable.HashMap[Id[Vehicle], RideHailAgentLocation]()
-  private val inServiceRideHailVehicles = mutable.HashMap[Id[Vehicle], RideHailAgentLocation]()
   private val pendingModifyPassengerScheduleAcks = mutable.HashMap[Int, RideHailResponse]()
   private var numPendingRoutingRequestsForReservations = 0
   private val parkingInquiryCache = collection.mutable.HashMap[Int, RideHailAgentLocation]()
@@ -291,11 +231,6 @@ class RideHailManager(
 
   // Are we in the middle of processing a batch?
   var currentlyProcessingTimeoutTrigger: Option[TriggerWithId] = None
-
-  //context.actorSelection("user/")
-  //rideHailIterationHistoryActor send message to ridheailiterationhsitoryactor
-
-  DebugLib.emptyFunctionForSettingBreakPoint()
 
   override def receive: Receive = LoggingReceive {
     case ev @ StopDrivingIfNoPassengerOnBoardReply(success, requestId, tick) =>
@@ -328,16 +263,16 @@ class RideHailManager(
         ) =>
       log.debug("RHM.NotifyVehicleResourceIdle: {}", ev)
       val vehicleId = vId.asInstanceOf[Id[Vehicle]]
-      val whenWhere = whenWhereOpt.getOrElse(getRideHailAgentLocation(vehicleId).currentLocationUTM)
+      val whenWhere = whenWhereOpt.getOrElse(vehicleManager.getRideHailAgentLocation(vehicleId).currentLocationUTM)
 
-      updateLocationOfAgent(vehicleId, whenWhere, getServiceStatusOf(vehicleId))
+      vehicleManager.updateLocationOfAgent(vehicleId, whenWhere, vehicleManager.getServiceStatusOf(vehicleId))
 
       //updateLocationOfAgent(vehicleId, whenWhereOpt, isAvailable = true)
       resources.get(agentsim.vehicleId2BeamVehicleId(vehicleId)).foreach { beamVehicle =>
         beamVehicle.driver.foreach { driver =>
           val rideHailAgentLocation =
             RideHailAgentLocation(driver, vehicleId, whenWhere)
-          vehicleState.put(vehicleId, beamVehicleState)
+          vehicleManager.vehicleState.put(vehicleId, beamVehicleState)
 
           if (modifyPassengerScheduleManager
                 .noPendingReservations(vehicleId) || modifyPassengerScheduleManager
@@ -361,17 +296,17 @@ class RideHailManager(
               // not enough range to make trip
 
               if (modifyPassengerScheduleManager.vehicleHasMoreThanOneOngoingRequests(vehicleId)) {
-                putOutOfService(rideHailAgentLocation)
+                vehicleManager.putOutOfService(rideHailAgentLocation)
                 sender() ! NotifyVehicleResourceIdleReply(triggerId, Vector[ScheduleTrigger]())
               } else {
                 log.debug("Not enough range: {}", vehicleId)
                 outOfServiceVehicleManager.registerTrigger(vehicleId, triggerId)
-                putOutOfService(rideHailAgentLocation)
+                vehicleManager.putOutOfService(rideHailAgentLocation)
                 findRefuelStationAndSendVehicle(rideHailAgentLocation)
               }
             } else {
               log.debug("Making available: {}", vehicleId)
-              makeAvailable(rideHailAgentLocation)
+              vehicleManager.makeAvailable(rideHailAgentLocation)
               // ridehail agent awaiting NotifyVehicleResourceIdleReply
               sender() ! NotifyVehicleResourceIdleReply(triggerId, Vector[ScheduleTrigger]())
             }
@@ -386,17 +321,17 @@ class RideHailManager(
 
     case NotifyResourceInUse(vId, whenWhere) =>
       val vehId = vId.asInstanceOf[Id[Vehicle]]
-      updateLocationOfAgent(vehId, whenWhere, InService)
+      vehicleManager.updateLocationOfAgent(vehId, whenWhere, InService)
 
     case BeamVehicleStateUpdate(id, beamVehicleState) =>
-      vehicleState.put(id, beamVehicleState)
+      vehicleManager.vehicleState.put(id, beamVehicleState)
 
     case MATSimNetwork(network) =>
       rideHailNetworkApi.setMATSimNetwork(network)
 
     case CheckInResource(vId, whenWhere) =>
       val vehicleId = vId.asInstanceOf[Id[Vehicle]]
-      updateLocationOfAgent(vehicleId, whenWhere.get, Available)
+      vehicleManager.updateLocationOfAgent(vehicleId, whenWhere.get, Available)
 
       if (whenWhere.get.time == 0) {
         resources(agentsim.vehicleId2BeamVehicleId(vehicleId)).driver
@@ -406,7 +341,7 @@ class RideHailManager(
             if (modifyPassengerScheduleManager
                   .noPendingReservations(vehicleId)) {
               log.debug("Checking in: {}", vehicleId)
-              makeAvailable(rideHailAgentLocation)
+              vehicleManager.makeAvailable(rideHailAgentLocation)
             }
             sender ! CheckInSuccess
             log.debug(
@@ -572,7 +507,7 @@ class RideHailManager(
       outOfServiceVehicleManager.initiateMovementToParkingDepot(vehicleId, passengerSchedule, tick)
 
     case RepositionVehicleRequest(passengerSchedule, tick, vehicleId, rideHailAgent) =>
-      if (getIdleVehicles.contains(vehicleId)) {
+      if (vehicleManager.getIdleVehicles.contains(vehicleId)) {
         modifyPassengerScheduleManager.repositionVehicle(
           passengerSchedule,
           tick,
@@ -708,93 +643,6 @@ class RideHailManager(
     }
   }
 
-  def getRideHailAgentLocation(vehicleId: Id[Vehicle]): RideHailAgentLocation = {
-    getServiceStatusOf(vehicleId) match {
-      case Available =>
-        availableRideHailVehicles(vehicleId)
-      case InService =>
-        inServiceRideHailVehicles(vehicleId)
-      case OutOfService =>
-        outOfServiceRideHailVehicles(vehicleId)
-    }
-  }
-
-  def getClosestIdleVehiclesWithinRadiusByETA(
-    pickupLocation: Coord,
-    radius: Double,
-    customerRequestTime: Long,
-    excludeRideHailVehicles: Set[Id[Vehicle]] = Set(),
-    secondsPerEuclideanMeterFactor: Double = 0.1 // (~13.4m/s)^-1 * 1.4
-  ): Vector[RideHailAgentETA] = {
-    var start = System.currentTimeMillis()
-    val nearbyAvailableRideHailAgents = availableRideHailAgentSpatialIndex
-      .getDisk(pickupLocation.getX, pickupLocation.getY, radius)
-      .asScala
-      .filter(x => availableRideHailVehicles.contains(x.vehicleId) && !excludeRideHailVehicles.contains(x.vehicleId))
-    var end = System.currentTimeMillis()
-    val diff1 = end - start
-
-    start = System.currentTimeMillis()
-    val times2RideHailAgents = nearbyAvailableRideHailAgents
-      .map { rideHailAgentLocation =>
-        val distance =
-          CoordUtils.calcProjectedEuclideanDistance(pickupLocation, rideHailAgentLocation.currentLocationUTM.loc)
-        // we consider the time to travel to the customer and the time before the vehicle is actually ready (due to
-        // already moving or dropping off a customer, etc.)
-        val extra = Math.max(rideHailAgentLocation.currentLocationUTM.time - customerRequestTime, 0)
-        val timeToCustomer = distance * secondsPerEuclideanMeterFactor + extra
-        RideHailAgentETA(rideHailAgentLocation, distance, timeToCustomer)
-      }
-      .toVector
-      .sortBy(_.timeToCustomer)
-    end = System.currentTimeMillis()
-    val diff2 = end - start
-
-    if (diff1 + diff2 > 100)
-      log.debug(
-        s"getClosestIdleVehiclesWithinRadiusByETA for $pickupLocation with $radius nearbyAvailableRideHailAgents: $diff1, diff2: $diff2. Total: ${diff1 + diff2} ms"
-      )
-
-    times2RideHailAgents
-  }
-
-  def getClosestIdleVehiclesWithinRadius(
-    pickupLocation: Coord,
-    radius: Double
-  ): Array[RideHailAgentLocation] = {
-    val idleVehicles = getIdleVehiclesWithinRadius(pickupLocation, radius).toArray
-    util.Arrays.sort(idleVehicles, RideHailAgentLocationWithRadiusOrdering)
-    idleVehicles.map { case (location, _) => location }
-  }
-
-  def getIdleVehiclesWithinRadius(
-    pickupLocation: Location,
-    radius: Double
-  ): Iterable[(RideHailAgentLocation, Double)] = {
-    val nearbyRideHailAgents = availableRideHailAgentSpatialIndex
-      .getDisk(pickupLocation.getX, pickupLocation.getY, radius)
-      .asScala
-      .view
-    val distances2RideHailAgents =
-      nearbyRideHailAgents.map(rideHailAgentLocation => {
-        val distance = CoordUtils
-          .calcProjectedEuclideanDistance(pickupLocation, rideHailAgentLocation.currentLocationUTM.loc)
-        (rideHailAgentLocation, distance)
-      })
-    distances2RideHailAgents.filter(x => availableRideHailVehicles.contains(x._1.vehicleId))
-  }
-
-  def getClosestIdleRideHailAgent(
-    pickupLocation: Coord,
-    radius: Double
-  ): Option[RideHailAgentLocation] = {
-    val idleVehicles = getIdleVehiclesWithinRadius(pickupLocation, radius)
-    if (idleVehicles.isEmpty) None
-    else {
-      val min = idleVehicles.min(RideHailAgentLocationWithRadiusOrdering)
-      Some(min._1)
-    }
-  }
 
   def attemptToCancelCurrentRideRequest(tick: Int, requestId: Int): Unit = {
     Option(travelProposalCache.getIfPresent(requestId.toString)) match {
@@ -813,9 +661,6 @@ class RideHailManager(
     }
 
   }
-
-  def getVehicleState(vehicleId: Id[Vehicle]): BeamVehicleState =
-    vehicleState(vehicleId)
 
   def createRoutingRequestsToCustomerAndDestination(
     requestTime: Int,
@@ -877,7 +722,7 @@ class RideHailManager(
     var numberOfTrips = 0
     for (repositionVehicle <- repositionVehicles) {
       val (vehicleId, destinationLocation) = repositionVehicle
-      val rideHailAgentLocation = getIdleVehicles(vehicleId)
+      val rideHailAgentLocation = vehicleManager.getIdleVehicles(vehicleId)
 
       sumOfDistances += beamServices.geo
         .distUTMInMeters(rideHailAgentLocation.currentLocationUTM.loc, destinationLocation)
@@ -889,145 +734,6 @@ class RideHailManager(
     DebugLib.emptyFunctionForSettingBreakPoint()
   }
 
-  def getIdleVehicles: mutable.HashMap[Id[Vehicle], RideHailAgentLocation] = {
-    availableRideHailVehicles
-  }
-
-  private def getServiceStatusOf(vehicleId: Id[Vehicle]): RideHailServiceStatus = {
-    if (availableRideHailVehicles.contains(vehicleId)) {
-      Available
-    } else if (inServiceRideHailVehicles.contains(vehicleId)) {
-      InService
-    } else if (outOfServiceRideHailVehicles.contains(vehicleId)) {
-      OutOfService
-    } else {
-      log.error(s"Vehicle {} does not have a service status, assuming out of service", vehicleId)
-      OutOfService
-    }
-  }
-
-  private def updateLocationOfAgent(
-    vehicleId: Id[Vehicle],
-    whenWhere: SpaceTime,
-    serviceStatus: RideHailServiceStatus
-  ) = {
-    serviceStatus match {
-      case Available =>
-        availableRideHailVehicles.get(vehicleId) match {
-          case Some(prevLocation) =>
-            val newLocation = prevLocation.copy(currentLocationUTM = whenWhere)
-            availableRideHailAgentSpatialIndex.remove(
-              prevLocation.currentLocationUTM.loc.getX,
-              prevLocation.currentLocationUTM.loc.getY,
-              prevLocation
-            )
-            availableRideHailAgentSpatialIndex.put(
-              newLocation.currentLocationUTM.loc.getX,
-              newLocation.currentLocationUTM.loc.getY,
-              newLocation
-            )
-            availableRideHailVehicles.put(newLocation.vehicleId, newLocation)
-          case None =>
-        }
-      case InService =>
-        inServiceRideHailVehicles.get(vehicleId) match {
-          case Some(prevLocation) =>
-            val newLocation = prevLocation.copy(currentLocationUTM = whenWhere)
-            inServiceRideHailAgentSpatialIndex.remove(
-              prevLocation.currentLocationUTM.loc.getX,
-              prevLocation.currentLocationUTM.loc.getY,
-              prevLocation
-            )
-            inServiceRideHailAgentSpatialIndex.put(
-              newLocation.currentLocationUTM.loc.getX,
-              newLocation.currentLocationUTM.loc.getY,
-              newLocation
-            )
-            inServiceRideHailVehicles.put(newLocation.vehicleId, newLocation)
-          case None =>
-        }
-      case OutOfService =>
-        outOfServiceRideHailVehicles.get(vehicleId) match {
-          case Some(prevLocation) =>
-            val newLocation = prevLocation.copy(currentLocationUTM = whenWhere)
-            outOfServiceRideHailAgentSpatialIndex.remove(
-              prevLocation.currentLocationUTM.loc.getX,
-              prevLocation.currentLocationUTM.loc.getY,
-              prevLocation
-            )
-            outOfServiceRideHailAgentSpatialIndex.put(
-              newLocation.currentLocationUTM.loc.getX,
-              newLocation.currentLocationUTM.loc.getY,
-              newLocation
-            )
-            outOfServiceRideHailVehicles.put(newLocation.vehicleId, newLocation)
-          case None =>
-        }
-    }
-  }
-
-  private def makeAvailable(agentLocation: RideHailAgentLocation) = {
-    availableRideHailVehicles.put(agentLocation.vehicleId, agentLocation)
-    availableRideHailAgentSpatialIndex.put(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    inServiceRideHailVehicles.remove(agentLocation.vehicleId)
-    inServiceRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    outOfServiceRideHailVehicles.remove(agentLocation.vehicleId)
-    outOfServiceRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-  }
-
-  private def putIntoService(agentLocation: RideHailAgentLocation) = {
-    availableRideHailVehicles.remove(agentLocation.vehicleId)
-    availableRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    outOfServiceRideHailVehicles.remove(agentLocation.vehicleId)
-    outOfServiceRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    inServiceRideHailVehicles.put(agentLocation.vehicleId, agentLocation)
-    inServiceRideHailAgentSpatialIndex.put(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-  }
-
-  private def putOutOfService(agentLocation: RideHailAgentLocation) = {
-    availableRideHailVehicles.remove(agentLocation.vehicleId)
-    availableRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    inServiceRideHailVehicles.remove(agentLocation.vehicleId)
-    inServiceRideHailAgentSpatialIndex.remove(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-    outOfServiceRideHailVehicles.put(agentLocation.vehicleId, agentLocation)
-    outOfServiceRideHailAgentSpatialIndex.put(
-      agentLocation.currentLocationUTM.loc.getX,
-      agentLocation.currentLocationUTM.loc.getY,
-      agentLocation
-    )
-  }
 
   private def handleReservation(request: RideHailRequest, tick: Int, travelProposal: TravelProposal): Unit = {
     surgePricingManager.addRideCost(
@@ -1037,7 +743,7 @@ class RideHailManager(
     )
 
     // This makes the vehicle unavailable for others to reserve
-    putIntoService(travelProposal.rideHailAgentLocation)
+    vehicleManager.putIntoService(travelProposal.rideHailAgentLocation)
 
     // Create confirmation info but stash until we receive ModifyPassengerScheduleAck
     pendingModifyPassengerScheduleAcks.put(
@@ -1222,8 +928,8 @@ class RideHailManager(
     }
 
     for ((vehicleId, destinationLocation) <- repositionVehicles) {
-      if (getIdleVehicles.contains(vehicleId)) {
-        val rideHailAgentLocation = getIdleVehicles(vehicleId)
+      if (vehicleManager.getIdleVehicles.contains(vehicleId)) {
+        val rideHailAgentLocation = vehicleManager.getIdleVehicles(vehicleId)
 
         val rideHailVehicleAtOrigin = StreetVehicle(
           rideHailAgentLocation.vehicleId,
