+++ src/main/scala/beam/agentsim/agents/ridehail/allocation/DebugRepositioning.scala
@@ -17,7 +17,7 @@ object DebugRepositioning {
 
       for (veh <- rideHailManager.resources.values) {
         spatialPlot.addPoint(
-          PointToPlot(rideHailManager.getRideHailAgentLocation(veh.id).currentLocationUTM.loc, Color.BLACK, 5)
+          PointToPlot(rideHailManager.vehicleManager.getRideHailAgentLocation(veh.id).currentLocationUTM.loc, Color.BLACK, 5)
         )
       }
