+++ src/main/scala/beam/agentsim/agents/ridehail/allocation/Pooling.scala
@@ -1,6 +1,7 @@
 package beam.agentsim.agents.ridehail.allocation
 
-import beam.agentsim.agents.ridehail.RideHailManager.{PoolingInfo, RideHailAgentLocation}
+import beam.agentsim.agents.ridehail.RideHailManager.{PoolingInfo}
+import beam.agentsim.agents.ridehail.RideHailVehicleManager.RideHailAgentLocation
 import beam.agentsim.agents.ridehail.{RideHailManager, RideHailRequest}
 import beam.agentsim.agents.vehicles.VehicleProtocol.StreetVehicle
 import beam.agentsim.events.SpaceTime
@@ -17,6 +18,7 @@ class Pooling(val rideHailManager: RideHailManager) extends RideHailResourceAllo
 
   override def respondToInquiry(inquiry: RideHailRequest): InquiryResponse = {
     rideHailManager
+        .vehicleManager
       .getClosestIdleVehiclesWithinRadiusByETA(
         inquiry.pickUpLocationUTM,
         rideHailManager.radiusInMeters,
@@ -54,7 +56,7 @@ class Pooling(val rideHailManager: RideHailManager) extends RideHailResourceAllo
       } else {
         // Make sure vehicle still available
         val vehicleId = routeResponses.head.itineraries.head.legs.head.beamVehicleId
-        if (rideHailManager.getIdleVehicles.contains(vehicleId) && !alreadyAllocated.contains(vehicleId)) {
+        if (rideHailManager.vehicleManager.getIdleVehicles.contains(vehicleId) && !alreadyAllocated.contains(vehicleId)) {
           alreadyAllocated = alreadyAllocated + vehicleId
           val pickDropIdAndLegs = routeResponses.map { rResp =>
             tempPickDropStore
@@ -64,7 +66,7 @@ class Pooling(val rideHailManager: RideHailManager) extends RideHailResourceAllo
           }
           allocResponses = allocResponses :+ VehicleMatchedToCustomers(
             request,
-            rideHailManager.getIdleVehicles(vehicleId),
+            rideHailManager.vehicleManager.getIdleVehicles(vehicleId),
             pickDropIdAndLegs
           )
         } else {
@@ -80,6 +82,7 @@ class Pooling(val rideHailManager: RideHailManager) extends RideHailResourceAllo
         case 1 =>
           val request = twoToPool.head
           rideHailManager
+              .vehicleManager
             .getClosestIdleVehiclesWithinRadiusByETA(
               request.pickUpLocationUTM,
               rideHailManager.radiusInMeters,
@@ -106,6 +109,7 @@ class Pooling(val rideHailManager: RideHailManager) extends RideHailResourceAllo
           val request2 = twoToPool.last
           val routingResponses2 = vehicleAllocationRequest.requests(request2)
           rideHailManager
+              .vehicleManager
             .getClosestIdleVehiclesWithinRadiusByETA(
               request1.pickUpLocationUTM,
               rideHailManager.radiusInMeters,
